fix retention; add test for adding old lines
