+++ src/internal/lokiutil/testloki/BUILD.bazel
@@ -60,6 +60,17 @@ go_test(
 go_test(
     name = "load_test",
     srcs = ["load_test.go"],
+    deps = [
+        ":testloki",
+        "//src/internal/lokiutil/client",
+        "//src/internal/pctx",
+        "//src/internal/require",
+    ],
+)
+
+go_test(
+    name = "parse_test",
+    srcs = ["parse_test.go"],
     embed = [":testloki"],
     deps = ["//src/internal/require"],
 )
